pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "api for JIA"
authors = ["Adam Elkobaitry <todoomz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
fastapi = "^0.111.0"
sqlmodel = "^0.0.19"
sqlalchemy = "^2.0.30"
black = "^24.4.2"
ruff = "^0.4.9"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pytest = "^8.2.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '/(\.git|\.venv|__pycache__)/'

[tool.ruff]
lint.select = ["E", "W", "F", "C90", "I", "D", "A", "B", "N", "YTT", "C4", "PT", "PTH", "ERA", "RUF"]
exclude = ['__pycache__', '.git', '.venv']
line-length = 120
lint.ignore = ['B008']

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pycodestyle]
max-doc-length = 100

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
]
addopts = [
    "-v",
]